Math/fixed: mark all conversions as "explicit"
Making the conversion operators explicit improves the awareness of
coders to those conversions.  Sometimes, implicit conversions have
side effects (e.g. information loss) or cause bloat.  Converting a
"fixed" to another data type should be an explicit decision.

This patch converts all remaining implicit conversions to explicit
ones, and fixes a few minor quirks.

+++ src/Device/Driver/AltairPro.cpp
@@ -250,7 +250,7 @@ AltairProDevice::PutTurnPoint(const TCHAR *propertyName, const Waypoint *waypoin
 
     CopyString(Name, waypoint->name.c_str(), ARRAY_SIZE(Name));
 
-    tmp = waypoint->location.latitude.Degrees();
+    tmp = (double)waypoint->location.latitude.Degrees();
 
     if(tmp < 0){
       NoS = 'S';
@@ -262,7 +262,7 @@ AltairProDevice::PutTurnPoint(const TCHAR *propertyName, const Waypoint *waypoin
     MinLat *= 60;
     MinLat *= 1000;
 
-    tmp = waypoint->location.longitude.Degrees();
+    tmp = (double)waypoint->location.longitude.Degrees();
 
     if (tmp < 0){
       EoW = 'W';
